Removed old science process
